refactor(terminal): extract PtySession into @kirby/terminal-pty
Splits the terminal lib so libs/terminal owns rendering + interfaces
only and the node-pty wrapper moves to its own backend lib. Sets up
the symmetric layout the upcoming tmux backend will mirror.

- libs/terminal-pty hosts PtySession + createPtyBackendFactory.
- libs/terminal drops the node-pty dependency and the PtySession
  re-export.
- pty-registry now constructs sessions via an injected
  SessionBackendFactory (defaulting to createPtyBackendFactory) and
  separates kill() (hard, used by killSession) from dispose()
  (soft, used by killAll).
- The PTY ↔ TerminalEmulator integration test moves with PtySession
  so libs/terminal's tests stay node-pty-free.
- hasAnySession() helper added for the upcoming Settings guard.
