+++ apps/cli/package.json
@@ -151,6 +151,7 @@
     "@inkjs/ui": "^2.0.0",
     "@kirby/logger": "*",
     "@kirby/terminal": "*",
+    "@kirby/terminal-pty": "*",
     "@kirby/vcs-azure-devops": "*",
     "@kirby/vcs-core": "*",
     "@kirby/vcs-github": "*",
