+++ apps/cli/tsconfig.app.json
@@ -34,6 +34,9 @@
       "path": "../../libs/vcs/azure-devops/tsconfig.lib.json"
     },
     {
+      "path": "../../libs/terminal-pty/tsconfig.lib.json"
+    },
+    {
       "path": "../../libs/terminal/tsconfig.lib.json"
     },
     {
