+++ libs/terminal-pty/package.json
@@ -22,6 +22,8 @@
     ]
   },
   "dependencies": {
+    "@kirby/terminal": "*",
+    "node-pty": "^1.1.0",
     "tslib": "^2.3.0"
   }
 }
