+++ libs/terminal-pty/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib/terminal-pty.js';
+export { PtySession, type PtySessionOptions } from './lib/pty-session.js';
+export { createPtyBackendFactory } from './lib/pty-backend.js';
