--- libs/terminal-pty/src/lib/terminal-pty.spec.ts
@@ -1,7 +1,0 @@
-import { terminalPty } from './terminal-pty.js';
-
-describe('terminalPty', () => {
-  it('should work', () => {
-    expect(terminalPty()).toEqual('terminal-pty');
-  });
-});
