--- libs/terminal-pty/src/lib/terminal-pty.ts
@@ -1,3 +1,0 @@
-export function terminalPty(): string {
-  return 'terminal-pty';
-}
