+++ libs/terminal/package.json
@@ -14,8 +14,7 @@
     "./package.json": "./package.json"
   },
   "dependencies": {
-    "@xterm/headless": "^6.0.0",
-    "node-pty": "^1.1.0"
+    "@xterm/headless": "^6.0.0"
   },
   "nx": {
     "name": "@kirby/terminal",
