+++ libs/terminal/src/index.ts
@@ -1,4 +1,3 @@
-export { PtySession, type PtySessionOptions } from './lib/pty-session.js';
 export {
   TerminalEmulator,
   type MouseTrackingMode,
